requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# RL & Environments
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Vision Models
transformers>=4.35.0
timm>=0.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Optional: Ego4D
# ego4d  # Install separately with: pip install ego4d

